Moved PCB files to submodule

+++ src/Motor_IMU_Shield.h
@@ -155,7 +155,7 @@ class IMU {
 				quaternion quaternionData;
 				vector linearAcceleration;
 				vector gravityVector;
-				int8_t temperature;
+				double temperature;
 			};
 			 void update(uint8_t dataType);
 		} data;
